Notebooks - Move MCS specific share helpers into folder
Prepars analyzing more areas (e.g. Cameron Pass)
