+++ notebooks/common.py
@@ -13,8 +13,6 @@ from snobedo.snotel import SnotelLocations
 
 from raster_file import RasterFile
 
-from nb_paths import *
-
 # Plot styles
 BOKEH_FONT = dict(
     fontsize={
@@ -75,4 +73,5 @@ def add_legend_box(ax, entries):
 def use_hvplot():
     import hvplot.xarray
     import hvplot.pandas
+    hv.extension('bokeh')
     pd.options.plotting.backend = 'holoviews'
